Feat: category shimmer ui 변경 #298 linearlayout을 2개 사용함으로 써 균등하게 보이게한 shimmer 사용
