+++ agijagi/app/src/main/java/likelion/project/agijagi/category/CategoryDetailInfoListFragment.kt
@@ -67,14 +67,16 @@ class CategoryDetailInfoListFragment : Fragment() {
             dataList.clear()
 
             ref.run {
-//                showTravelData(true)
+                showTravelData(true)
                 if (category == "All" && is_custom == 0) {
 
                     get()
                         .addOnSuccessListener { documents ->
                             val filteredDocuments =
                                 documents.filter { it.getString("state") != "숨김" }
+
                             setRecyclerCategoryList(filteredDocuments)
+                            showTravelData(false)
 
                         }.addOnFailureListener {
                             Log.d("category", "데이터")
@@ -86,7 +88,9 @@ class CategoryDetailInfoListFragment : Fragment() {
                         .addOnSuccessListener { documents ->
                             val filteredDocuments =
                                 documents.filter { it.getString("state") != "숨김" }
+
                             setRecyclerCategoryList(filteredDocuments)
+                            showTravelData(false)
 
                         }.addOnFailureListener {
                             Log.d("category", "데이터")
@@ -99,7 +103,9 @@ class CategoryDetailInfoListFragment : Fragment() {
                         .addOnSuccessListener { documents ->
                             val filteredDocuments =
                                 documents.filter { it.getString("state") != "숨김" }
+
                             setRecyclerCategoryList(filteredDocuments)
+                            showTravelData(false)
 
                         }.addOnFailureListener {
                             Log.d("category", "데이터")
@@ -112,13 +118,15 @@ class CategoryDetailInfoListFragment : Fragment() {
                         .addOnSuccessListener { documents ->
                             val filteredDocuments =
                                 documents.filter { it.getString("state") != "숨김" }
+
                             setRecyclerCategoryList(filteredDocuments)
+                            showTravelData(false)
 
                         }.addOnFailureListener {
                             Log.d("category", "데이터")
                         }
                 } else { }
-                showTravelData(false)
+
             }
         }
