+++ agijagi/app/src/main/java/likelion/project/agijagi/productoption/product/CustomProductDetailFragment.kt
@@ -1,11 +1,11 @@
-package likelion.project.agijagi.product
+package likelion.project.agijagi.productoption.product
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
+import androidx.core.content.ContextCompat
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.navigation.findNavController
@@ -16,15 +16,20 @@ import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.ktx.storage
 import io.supercharge.shimmerlayout.ShimmerLayout
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.cancel
+import kotlinx.coroutines.launch
 import likelion.project.agijagi.MainActivity.Companion.displayDialogUserNotLogin
 import likelion.project.agijagi.R
-import likelion.project.agijagi.databinding.FragmentProductDetailBinding
+import likelion.project.agijagi.databinding.FragmentCustomProductDetailBinding
 import likelion.project.agijagi.model.UserModel
 import java.text.DecimalFormat
 
-class ProductDetailFragment : Fragment() {
+class CustomProductDetailFragment : Fragment() {
 
-    private var _binding: FragmentProductDetailBinding? = null
+    private var _binding: FragmentCustomProductDetailBinding? = null
     private val binding get() = _binding!!
 
     val dec = DecimalFormat("#,###")
@@ -36,7 +41,7 @@ class ProductDetailFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentProductDetailBinding.inflate(inflater)
+        _binding = FragmentCustomProductDetailBinding.inflate(inflater)
 
         return binding.root
     }
@@ -44,37 +49,45 @@ class ProductDetailFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val productId = getProductId()
+        val productId = getCustomProductId()
 
-        loadProductDataAndInitViews(productId)
+        loadCustomProductDataAndInitViews(productId)
+        setupFloorPlanDownloadButton()
         setupFavoriteButton(productId)
+
     }
 
-    private fun getProductId(): String {
+    private fun getCustomProductId(): String {
         return arguments?.getString("prodId").toString()
     }
 
-    private fun loadProductDataAndInitViews(productId: String) {
+    private fun loadCustomProductDataAndInitViews(productId: String) {
         binding.run {
+
+
             val shimmerLayoutImages = listOf(
-                shimmerLayoutProductDetailImage1,
-                shimmerLayoutProductDetailImage2,
-                shimmerLayoutProductDetailImage3,
-                shimmerLayoutProductDetailImage4,
-                shimmerLayoutProductDetailImage5,
-                shimmerLayoutProductDetailImage6
+                shimmerLayoutCustomProductDetailImage1,
+                shimmerLayoutCustomProductDetailImage2,
+                shimmerLayoutCustomProductDetailImage3,
+                shimmerLayoutCustomProductDetailImage4,
+                shimmerLayoutCustomProductDetailImage5,
+                shimmerLayoutCustomProductDetailImage6
             )
 
             val imageViews = listOf(
-                imageviewProductDetailImage1,
-                imageviewProductDetailImage2,
-                imageviewProductDetailImage3,
-                imageviewProductDetailImage4,
-                imageviewProductDetailImage5,
-                imageviewProductDetailImage6
+                imageviewCustomProductDetailImage1,
+                imageviewCustomProductDetailImage2,
+                imageviewCustomProductDetailImage3,
+                imageviewCustomProductDetailImage4,
+                imageviewCustomProductDetailImage5,
+                imageviewCustomProductDetailImage6
             )
 
-            startShimmerAnimations(shimmerLayoutProductDetailThumbnailImage, shimmerLayoutImages)
+            startShimmerAnimations(
+                shimmerLayoutCustomProductDetailThumbnailImage,
+                shimmerLayoutImages
+            )
+
 
             db.collection("product").document(productId).get().addOnSuccessListener {
                 val thumbnailImage = it["thumbnail_image"].toString()
@@ -86,11 +99,11 @@ class ProductDetailFragment : Fragment() {
                 val state = it["state"].toString()
 
                 storageRef.child(thumbnailImage).downloadUrl.addOnSuccessListener { thumbnailUri ->
-                    shimmerLayoutProductDetailThumbnailImage.stopShimmerAnimation()
-                    Glide.with(this@ProductDetailFragment)
+                    shimmerLayoutCustomProductDetailThumbnailImage.stopShimmerAnimation()
+                    Glide.with(this@CustomProductDetailFragment)
                         .load(thumbnailUri)
                         .placeholder(R.drawable.product_detail_default_image)
-                        .into(imageviewProductDetailThumbnailImage)
+                        .into(imageviewCustomProductDetailThumbnailImage)
 
                     displayProductInfo(brand, name, price, name, detail)
 
@@ -118,11 +131,11 @@ class ProductDetailFragment : Fragment() {
         detail: String
     ) {
         binding.run {
-            textviewProductDetailBrand.text = brand
-            textviewProductDetailName.text = name
-            textviewProductDetailPrice.text = price
-            textviewProductDetailInfoTitle.text = title
-            textviewProductDetailInfoDescription.text = detail
+            textviewCustomProductDetailBrand.text = brand
+            textviewCustomProductDetailName.text = name
+            textviewCustomProductDetailPrice.text = price
+            textviewCustomProductDetailInfoTitle.text = title
+            textviewCustomProductDetailInfoDescription.text = detail
         }
     }
 
@@ -137,13 +150,14 @@ class ProductDetailFragment : Fragment() {
             storageRef.child(image[idx].toString()).downloadUrl.addOnSuccessListener { imageUri ->
                 shimmerLayoutImages[idx].stopShimmerAnimation()
 
-                Glide.with(this@ProductDetailFragment)
+                Glide.with(this@CustomProductDetailFragment)
                     .load(imageUri)
                     .placeholder(R.drawable.product_detail_default_image)
                     .into(imageViews[idx])
 
-                setupToolbar()
                 setupPurchaseButton(productId, state)
+                setupToolbar()
+                setupFloatingButton()
             }
         }
 
@@ -153,7 +167,7 @@ class ProductDetailFragment : Fragment() {
     }
 
     private fun setupToolbar() {
-        binding.toolbarProductDetail.run {
+        binding.toolbarCustomProductDetail.run {
             setNavigationOnClickListener {
                 findNavController().popBackStack()
             }
@@ -162,12 +176,12 @@ class ProductDetailFragment : Fragment() {
                     displayDialogUserNotLogin(
                         requireContext(),
                         findNavController(),
-                        R.id.action_productDetailFragment_to_loginFragment
+                        R.id.action_customProductDetailFragment_to_loginFragment
                     )
                 } else {
                     when (it.itemId) {
                         R.id.menu_product_detail_shopping -> {
-                            findNavController().navigate(R.id.action_productDetailFragment_to_shoppingListFragment)
+                            findNavController().navigate(R.id.action_customProductDetailFragment_to_shoppingListFragment)
                         }
                     }
                 }
@@ -176,9 +190,30 @@ class ProductDetailFragment : Fragment() {
         }
     }
 
+    private fun setupFloatingButton() {
+        binding.customFloatingButtonCustomProductDetailToChatting.customFloatingButtonLayout.setOnClickListener {
+            if (UserModel.uid == "") {
+                displayDialogUserNotLogin(
+                    requireContext(),
+                    findNavController(),
+                    R.id.action_customProductDetailFragment_to_loginFragment
+                )
+            } else {
+                findNavController()
+                    .navigate(R.id.action_customProductDetailFragment_to_chattingListFragment)
+            }
+        }
+    }
+
+    private fun setupFloorPlanDownloadButton() {
+        binding.buttonCustomProductDetailDownloadFloorPlan.setOnClickListener {
+            Snackbar.make(it, "도면 다운로드가 완료되었습니다.", Snackbar.LENGTH_SHORT).show()
+        }
+    }
+
     private fun setupFavoriteButton(productId: String) {
         val buyerId = UserModel.roleId
-        binding.imageButtonProductDetailFavorite.run {
+        binding.imageButtonCustomProductDetailFavorite.run {
             if (UserModel.uid != "") {
                 db.collection("buyer")
                     .document(buyerId)
@@ -198,17 +233,17 @@ class ProductDetailFragment : Fragment() {
                     displayDialogUserNotLogin(
                         requireContext(),
                         findNavController(),
-                        R.id.action_productDetailFragment_to_loginFragment
+                        R.id.action_customProductDetailFragment_to_loginFragment
                     )
                 } else {
                     it.isSelected = it.isSelected != true
                     if (it.isSelected) {
-                        val prodId = hashMapOf("prodId" to productId)
+                        val product = hashMapOf("prodId" to productId)
                         db.collection("buyer")
                             .document(buyerId)
                             .collection("wish")
                             .document(productId)
-                            .set(prodId)
+                            .set(product)
                     } else {
                         db.collection("buyer")
                             .document(buyerId)
@@ -222,9 +257,10 @@ class ProductDetailFragment : Fragment() {
     }
 
     private fun setupPurchaseButton(productId: String, state: String) {
-        binding.buttonProductDetailPurchase.run {
+        binding.buttonCustomProductDetailPurchase.run {
             if (state == "품절") {
                 setBackgroundResource(R.drawable.wide_box_rounded_purchase_button_inactive)
+                setTextColor(ContextCompat.getColor(context, R.color.jagi_black_42))
                 text = "품절"
             }
             setOnClickListener {
@@ -234,12 +270,15 @@ class ProductDetailFragment : Fragment() {
                     displayDialogUserNotLogin(
                         requireContext(),
                         findNavController(),
-                        R.id.action_productDetailFragment_to_loginFragment
+                        R.id.action_customProductDetailFragment_to_loginFragment
                     )
                 } else {
                     val bundle = bundleOf("prodId" to productId)
                     it.findNavController()
-                        .navigate(R.id.action_productDetailFragment_to_readyMadeOptionFragment, bundle)
+                        .navigate(
+                            R.id.action_customProductDetailFragment_to_customOptionFragment,
+                            bundle
+                        )
                 }
             }
         }
@@ -251,4 +290,3 @@ class ProductDetailFragment : Fragment() {
     }
 
 }
-
