+++ agijagi/app/src/main/res/layout/fragment_custom_product_detail.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".product.CustomProductDetailFragment">
+    tools:context=".productoption.product.CustomProductDetailFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
