+++ agijagi/app/src/main/res/layout/item_category_list_shimmer.xml
@@ -1,57 +1,127 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:paddingBottom="16dp"
-    android:orientation="vertical"
+    android:orientation="horizontal"
     android:paddingHorizontal="4dp">
 
-    <com.google.android.material.card.MaterialCardView
-        android:id="@+id/cardview_category_list_shimmer"
-        android:layout_width="wrap_content"
-        android:layout_height="200dp"
-        app:cardCornerRadius="8dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:strokeColor="@color/jagi_brown">
+    <LinearLayout
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:padding="4dp">
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
 
-        <ImageView
-            android:id="@+id/imageview_category_list_image_shimmer"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:adjustViewBounds="true"
-            android:scaleType="centerCrop"
-            android:src="@drawable/category_list_item_default_image" />
+            <com.google.android.material.card.MaterialCardView
+                android:id="@+id/cardview_category_list_shimmer_one"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:cardCornerRadius="8dp"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:strokeColor="@color/jagi_brown">
 
-    </com.google.android.material.card.MaterialCardView>
+                <ImageView
+                    android:id="@+id/imageview_category_list_image_shimmer_one"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:adjustViewBounds="true"
+                    android:scaleType="centerCrop"
+                    android:src="@drawable/category_list_item_default_image" />
 
-    <TextView
-        android:id="@+id/textview_category_list_brand"
-        android:layout_width="56dp"
-        android:layout_height="16dp"
-        android:layout_marginTop="8dp"
-        android:background="#9F9F9F"
-        android:textColor="@color/brand_name"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/cardview_category_list_shimmer" />
+            </com.google.android.material.card.MaterialCardView>
 
-    <TextView
-        android:id="@+id/textview_category_list_name"
-        android:layout_width="80dp"
-        android:layout_height="16dp"
-        android:layout_marginTop="4dp"
-        android:background="#9F9F9F"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/textview_category_list_brand" />
+            <TextView
+                android:id="@+id/textview_category_list_brand_one"
+                android:layout_width="56dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="8dp"
+                android:background="#9F9F9F"
+                android:textColor="@color/brand_name"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/cardview_category_list_shimmer_one" />
 
-    <TextView
-        android:id="@+id/textview_category_list_price"
-        android:layout_width="68dp"
-        android:layout_height="16dp"
-        android:layout_marginTop="0.5dp"
-        android:background="#9F9F9F"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/textview_category_list_name" />
+            <TextView
+                android:id="@+id/textview_category_list_name_one"
+                android:layout_width="80dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="4dp"
+                android:background="#9F9F9F"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/textview_category_list_brand_one" />
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+            <TextView
+                android:id="@+id/textview_category_list_price_one"
+                android:layout_width="68dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="0.5dp"
+                android:background="#9F9F9F"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/textview_category_list_name_one" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:padding="4dp">
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content">
+
+            <com.google.android.material.card.MaterialCardView
+                android:id="@+id/cardview_category_list_shimmer_two"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                app:cardCornerRadius="8dp"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:strokeColor="@color/jagi_brown">
+
+                <ImageView
+                    android:id="@+id/imageview_category_list_image_shimmer_two"
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:adjustViewBounds="true"
+                    android:scaleType="centerCrop"
+                    android:src="@drawable/category_list_item_default_image" />
+
+            </com.google.android.material.card.MaterialCardView>
+
+            <TextView
+                android:id="@+id/textview_category_list_brand_two"
+                android:layout_width="56dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="8dp"
+                android:background="#9F9F9F"
+                android:textColor="@color/brand_name"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/cardview_category_list_shimmer_two" />
+
+            <TextView
+                android:id="@+id/textview_category_list_name_two"
+                android:layout_width="80dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="4dp"
+                android:background="#9F9F9F"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/textview_category_list_brand_two" />
+
+            <TextView
+                android:id="@+id/textview_category_list_price_two"
+                android:layout_width="68dp"
+                android:layout_height="16dp"
+                android:layout_marginTop="0.5dp"
+                android:background="#9F9F9F"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/textview_category_list_name_two" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </LinearLayout>
+
+</LinearLayout>
