+++ agijagi/app/src/main/res/navigation/navigation_graph.xml
@@ -96,7 +96,7 @@
 
     <fragment
         android:id="@+id/customProductDetailFragment"
-        android:name="likelion.project.agijagi.product.CustomProductDetailFragment"
+        android:name="likelion.project.agijagi.productoption.product.CustomProductDetailFragment"
         android:label="fragment_custom_product_detail">
         <action
             android:id="@+id/action_customProductDetailFragment_to_customOptionFragment"
@@ -132,7 +132,7 @@
 
     <fragment
         android:id="@+id/productDetailFragment"
-        android:name="likelion.project.agijagi.product.ProductDetailFragment"
+        android:name="likelion.project.agijagi.productoption.product.ProductDetailFragment"
         android:label="fragment_product_detail">
         <action
             android:id="@+id/action_productDetailFragment_to_readyMadeOptionFragment"
